+++ config-manager-proc/src/utils/config.rs
@@ -1,15 +1,13 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) 2022 JSRPC “Kryptonite”
 
-// TODO
 use std::collections::HashSet;
-
-use strum::IntoEnumIterator;
 
 use super::attributes::*;
 use crate::*;
 
-fn str_to_config_format_repr(s: &str) -> String {
+fn str_to_config_format_repr<S: AsRef<str>>(s: S, span: Span) -> Result<String> {
+    let s = s.as_ref().trim_matches('"');
     match s {
         "json" | "json5" | "toml" | "yaml" | "ron" => {
             let capitalize_first = |s: &str| -> String {
@@ -20,13 +18,14 @@ fn str_to_config_format_repr(s: &str) -> String {
 
             let accepted_format = capitalize_first(s);
             let pref = "::config_manager::__private::config::FileFormat::".to_string();
-            pref + &accepted_format
+            Ok(pref + &accepted_format)
         }
-        _ => panic!("{s} format is not supported"),
+        _ => panic_span!(span, "{s} format is not supported"),
     }
 }
 
 struct ParsedConfigFileAttributes {
+    span: Span,
     file_format: String,
     clap_info: Option<NormalClapFieldInfo>,
     env_key: Option<String>,
@@ -45,89 +44,60 @@ fn handle_file_attributes(class_attributes: &[Meta]) -> Result<Vec<ParsedConfigF
 fn handle_file_attribute(attr: &Meta) -> Result<ParsedConfigFileAttributes> {
     let nested = attr
         .require_list()?
-        .parse_args_with(Punctuated::<Meta, Token![,]>::parse_terminated)
-        .expect(
-            "config arguments must match file(format = \"...\", clap_key = \
-                             \"...\", env = \"...\", default = \"...\", optional = \
-                             true/[false])",
-        );
+        .parse_args_with(Punctuated::<Meta, Token![,]>::parse_terminated)?;
 
     let mut clap_info = None;
-    let mut config_file_attributes: Vec<_> = ConfigFileAttr::iter()
-        .map(|ty| OptionalAttribute {
-            value: None,
-            accepted_literals: match ty {
-                ConfigFileAttr::Optional => AcceptedLiterals::Bool,
-                _ => AcceptedLiterals::String,
-            },
-            ty,
-        })
-        .collect();
-
-    'next_arg: for arg in nested {
-        if path_to_string(arg.path()) == CLAP_KEY {
-            let clap_list = arg
-                .require_list()
-                .expect("clap attribute must match \"clap(...)\"");
-
-            clap_info = Some(parse_clap_field_attribute(clap_list, false)?);
-            continue 'next_arg;
-        } else {
-            for attr in &mut config_file_attributes {
-                match try_set_optional_attribute::<ConfigFileAttr>(&arg, attr, false) {
-                    SetOptionalAttrResult::NameMismatch => (),
-                    SetOptionalAttrResult::Set => continue 'next_arg,
-                    SetOptionalAttrResult::ErrorAlreadySet => {
-                        panic!("attempted to set {} twice", attr.ty)
-                    }
-                }
-            }
-            panic!("unknown attribute: {:#?}", arg)
-        }
-    }
-
     let mut file_format = None;
     let mut env_key = None;
     let mut optional = false;
     let mut default = None;
 
-    for attr in config_file_attributes {
-        match attr.ty {
-            ConfigFileAttr::EnvKey => env_key = attr.value,
-            ConfigFileAttr::Optional => {
-                if let Some(attr) = attr.value {
-                    optional = attr.parse().unwrap()
+    for arg in nested {
+        match path_to_string(arg.path()).as_str() {
+            "clap" => {
+                if clap_info.is_some() {
+                    panic_span!(arg.span(), "attempted to set clap twice")
                 }
-            }
-            ConfigFileAttr::Default => {
-                if let Some(attr) = attr.value {
-                    default = Some(attr)
-                }
-            }
-            ConfigFileAttr::Format => {
-                if let Some(format_attr) = attr.value {
-                    assert_eq!(format_attr.chars().next().unwrap(), '"');
-                    assert_eq!(format_attr.chars().last().unwrap(), '"');
+                let clap_list = arg.require_list().map_err(|_err| {
+                    Error::new(arg.span(), "clap attribute must match \"clap(...)\"")
+                })?;
 
-                    let drop_fst_and_lst: String = format_attr
-                        .chars()
-                        .skip(1)
-                        .take(format_attr.len() - 2)
-                        .collect();
-                    file_format = Some(str_to_config_format_repr(&drop_fst_and_lst))
-                }
+                clap_info = Some(parse_clap_field_attribute(clap_list, false)?);
             }
+            "format" => {
+                let file_attr = set_config_attr(file_format.is_some(), &arg, "format")?.unwrap();
+                file_format = Some(str_to_config_format_repr(file_attr, arg.span())?);
+            }
+            "env" => env_key = set_config_attr(env_key.is_some(), &arg, "env")?,
+            "optional" => {
+                if optional {
+                    panic_span!(arg.span(), "attempted to set optional twice")
+                }
+                if !matches!(arg, Meta::Path(_)) {
+                    panic_span!(
+                        arg.span(),
+                        "optional cannot take values. Usage: file(optional)"
+                    )
+                }
+                optional = true;
+            }
+            "default" => default = set_config_attr(default.is_some(), &arg, "default")?,
+            other => panic_span!(arg.span(), "unknown attribute {other}"),
         }
     }
 
     if clap_info.is_none() && env_key.is_none() && default.is_none() {
-        panic!("you must specify at least one of (clap, env, default)");
+        panic_span!(
+            attr.span(),
+            "you must specify at least one of (clap, env, default)"
+        );
     }
     if let Some(clap_info) = &clap_info {
         if let ClapOption::None | ClapOption::Empty = clap_info.long {
-            panic!(
-                "if #[clap] attribute is specified for configuration file, nested \
-                             `long = ...` must be provided"
+            panic_span!(
+                clap_info.span,
+                "if #[clap] attribute is specified for configuration file, nested `long = ...` \
+                    must be provided. Otherwise it's impossible to determine arg name"
             );
         }
     }
@@ -137,13 +107,29 @@ fn handle_file_attribute(attr: &Meta) -> Result<ParsedConfigFileAttributes> {
         .transpose()?;
 
     Ok(ParsedConfigFileAttributes {
+        span: attr.span(),
         default,
         optional,
         clap_info,
         env_key,
-        file_format: file_format
-            .unwrap_or_else(|| panic!("`format` attribute of config file must be set")),
+        file_format: file_format.ok_or_else(|| {
+            Error::new(attr.span(), "`format` attribute of config file must be set")
+        })?,
     })
+}
+
+fn set_config_attr<N: AsRef<str>>(
+    already_set: bool,
+    arg: &Meta,
+    attr_name: N,
+) -> Result<Option<String>> {
+    let attr_name = attr_name.as_ref();
+    if already_set {
+        panic_span!(arg.span(), "attempted to set {attr_name} twice")
+    }
+    Ok(Some(meta_to_option(arg).ok_or_else(|| {
+        Error::new(arg.span(), format!("file({attr_name}) can't be empty"))
+    })?))
 }
 
 pub(crate) struct ConfigFilesInfo {
@@ -166,6 +152,7 @@ pub(crate) fn extract_configs_info(class_attributes: &[Meta]) -> Result<ConfigFi
     let mut config_env_keys = HashSet::<String>::new();
 
     for ParsedConfigFileAttributes {
+        span,
         file_format,
         clap_info,
         env_key,
@@ -174,12 +161,10 @@ pub(crate) fn extract_configs_info(class_attributes: &[Meta]) -> Result<ConfigFi
     } in handle_file_attributes(class_attributes)?
     {
         if optional && default.is_some() {
-            panic!(
-                concat!(
-                    "setting optional to true while specifying the default path is pointless, ",
-                    "since the config file will always be tried to parse [default = {}]"
-                ),
-                default.unwrap()
+            panic_span!(
+                span,
+                "setting optional to true while specifying the default path is pointless, \
+                    since the config file will always be tried to parse"
             );
         }
 
@@ -187,7 +172,10 @@ pub(crate) fn extract_configs_info(class_attributes: &[Meta]) -> Result<ConfigFi
             let clap_long = clap_info.long.clone();
             let is_new = config_clap_keys.insert(clap_long.clone());
             if !is_new {
-                panic!("config file with clap key {clap_long} already specified");
+                panic_span!(
+                    span,
+                    "config file with clap key {clap_long} already specified"
+                );
             }
             TokenStream::from_str(&format!("::std::option::Option::Some({clap_long})"))
         } else {
@@ -198,7 +186,7 @@ pub(crate) fn extract_configs_info(class_attributes: &[Meta]) -> Result<ConfigFi
         let env_key = if let Some(env_key) = env_key {
             let is_new = config_env_keys.insert(env_key.clone());
             if !is_new {
-                panic!("config file with env key {env_key} already specified");
+                panic_span!(span, "config file with env key {env_key} already specified");
             }
             TokenStream::from_str(&format!("::std::option::Option::Some({env_key})"))
         } else {