+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] - ReleaseDate
+### Changed
+- `#config[file(optional)]` attribute now does not take value. File is required when not mentioned and optional when this attribute is set.
+
 ## [0.3.1](https://github.com/3xMike/config-manager/releases/tag/0.3.1) - 2025-03-12
 ### Added
 - Allow to use clap(help), clap(long_help) and clap(long_about) without value.